# Reverse entry iterator pool