The `iter.NewReversedIter` was causing a lot of allocs. This uses a pool and remove the labels string since we never needed it.

benchcmp:
```
benchmark                        old ns/op     new ns/op     delta
BenchmarkBackwardIterator-16     15943073      14537418      -8.82%

benchmark                        old allocs     new allocs     delta
BenchmarkBackwardIterator-16     47217          47225          +0.02%

benchmark                        old bytes     new bytes     delta
BenchmarkBackwardIterator-16     27444423      23664350      -13.77%
```