@cyriltovena 2020-05-11
> Overall, lgtm once you get your tests passing. I'm a little wary of having implementations which only satisfy the `EntryIterator` ifc in compilation, but not in reality (returning `""` for `Labels()`). That being said, I see this isn't the only place we do this, so I don't want to block it.

I wasn't sure how to do this properly, I just catched that with labels it's a bit slower, not much but with the amount of those struct allocated it can make a difference.